feat: enhance GoogleOAuthButton & Input components with additional props
