+++ apps/client/src/app/[lng]/(auth)/login/page.tsx
@@ -61,7 +61,14 @@ const Page = () => {
         linkLabel: t('auth.login.buttons.register')
       }}
     >
-      <GoogleOAuthButton onSuccess={onGoogleLogin} />
+      <GoogleOAuthButton
+        onSuccess={onGoogleLogin}
+        type="button"
+        variant="outlined"
+        color="primary"
+        width="full"
+        size="xl"
+      />
       <div className="flex items-center my-4">
         <hr className="flex-grow border-gray-300" />
         <span className="px-3 text-gray-500 text-sm">{t('auth.or')}</span>
@@ -72,7 +79,7 @@ const Page = () => {
         <Input
           label={{ text: t('auth.email') }}
           input={{ id: 'email', placeholder: t('auth.enterEmail'), ...register('email', { required: true }) }}
-          size="lg"
+          size="xl"
           error={errors.email?.message}
         />
         <Password
@@ -83,8 +90,9 @@ const Page = () => {
             ...register('password', { required: true })
           }}
           error={errors.password?.message}
+          size="xl"
         />
-        <Button width="full" type="submit" loading={isPending}>
+        <Button width="full" type="submit" loading={isPending} size="xl">
           {t('auth.login.buttons.login')}
         </Button>
       </form>
