+++ apps/client/src/components/Password/Password.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Input } from '@client/components';
-import { InputProps } from '@client/types';
+import { InputProps } from '@client/utils';
 import { faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
