--- apps/client/src/types/input.ts
@@ -1,15 +1,0 @@
-import { ImageProps } from 'next/image';
-
-type LabelProps = {
-  text: React.ReactNode | string;
-  position?: 'left' | 'right' | 'bottom';
-} & React.LabelHTMLAttributes<HTMLLabelElement>;
-
-export interface InputProps {
-  label?: LabelProps;
-  icon?: ImageProps | React.ReactElement;
-  input?: React.InputHTMLAttributes<HTMLInputElement>;
-  iconPosition?: 'left' | 'right';
-  size?: 'sm' | 'md' | 'lg' | 'xl';
-  error?: string | boolean;
-}
